Support more hardware features
Originally, wiphy can only simulated 802.11 h/w with 2GHz band in 6th
channel, and rates in 1/2/5.5/11 Mbps, which may limit someone who wants
to configure with further PHY like 802.11g.

So in this commit, I extend the wiphy with the ability to simulate 1st to 14th
channels in 2GHz band, and rates in 1/2/5.5/11/6/9/12/18/24/36/48/54 Mbps.