Support passive scan
Two changes were made in this commit:
(1) Passive scanning of vwifi: Once we execute hostapd to start AP mode
, it will initialize an htrimer and set its countdown time to the
beacon interval input by the user application. When the time is up, the
callback function will be triggered, and it will Pass the AP's beacon
to other stations.
(2) Supports 5 GHZ band: The frequency is calculated as
5 GHz + (5 * channel index).

Finally, some errors in vwifi will be corrected so that verify.sh can
be executed correctly.