Use spinlock when accessing vif_list
In this commit, access to vif_list is protected by spinlock.
In the process context, considering the possibility of being
interrupted by softirq (timer callback function), the reading and
operation of vif_list are protected with spin_lock_bh(), and the
operation of vif_list in the timer callback function is protected
with spin_lock().