Introduce a userspace tool (#53)
This commit introduces a userspace tool, vwifi-tool, which can
display the status of the driver and allows users to specify a
blocklist.

This tool will show the status of the vwifi driver by reading
from /sys/module/vwifi/initstate. Users can only set a blocklist
for the vwifi driver when it is loaded.

Users can run vwifi-tool with command-line options, parsed by
getopt(), to set or unset a user-specific blocklist pair. The
blocklist will be copied and sent to the vwifi driver via a
netlink socket.

Close #48