Enforce consistent naming scheme
Changes are made to the prefix of functions, variables and related
 symbols from "owl" to "vwifi", just like the name of this project.
At the same time, the prefix of the name for virtual interfaces are
changed, too. Virtual interfaces used "owl" as the prefix of their
name but now they use "vw" for the prefix of their name.
For example "owl0" are changed to "vw0".

Close #52