Implement customized naming for wiphy
The default phy%d naming for wiphy in linux kernel depends on the value
 of a static variable wiphy_counter. Since they never attempts to
 decrease the value of wiphy_counter even a wiphy is unregistered or
 freed, this behavior ensures the naming and indexing for wiphy will
 be absolutely unique.

However, the kernel might have other projects also utilize wiphy
structure, which will cause some confusion of wiphy's index and naming
when using `struct wiphy`. We implement a custom-made name "vw_phy%d"
for wiphy in our project, in order to seperate the naming and indexing
for `struct wiphy` in our project.

As for the troublesome behavior causing by the never-decreasing index of
wiphy structure as decribed in #54, we suggest to make some changes when
doing manual testing by finding the wiphy index of the interface first,
and use the index to query the actual wiphy's name for the interface as
describe in the changes of README.md.

Close #54