Modify indexing for function vwifi_interface_add
Since we implement a counter for wiphy in our project, which is
vwifi_wiphy_counter. The relationship between virtual interfaces and
wiphy is always one-to-one, so the indexing for virtual interfaces can
be the same as wiphy.

Originally, the index for virtual interface is passed as a function
parameters into the function vwifi_interface_add(). Now we modify the
function, instead of using parameters as index for virtual interface,
we use the value of vwifi_wiphy_counter to keep a better consistency.